+++ src/shared/data/profile.js
@@ -41,7 +41,6 @@ const profile = {
       "steam",
       "backloggd",
       "letterboxd",
-      "notion",
     ],
   },