+++ src/shared/i18n/locales/ja/index.js
@@ -337,7 +337,6 @@ export default {
       steam: "Steam",
       backloggd: "Backloggd",
       letterboxd: "Letterboxd",
-      notion: "Notion",
     },
   },
   letterboxd: {