+++ src/shared/i18n/locales/ru/index.js
@@ -338,7 +338,6 @@ export default {
       steam: "Steam",
       backloggd: "Backloggd",
       letterboxd: "Letterboxd",
-      notion: "Notion",
     },
   },
   letterboxd: {